==> code/r/00-Update-and-install-packages.R <==
# Update R
install.packages("installr")

library(installr)

updateR()

#Add in packages
install.packages("dplyr")
install.packages("tidyr")
install.packages("ggplot2")
install.packages("forcats")
install.packages("lubridate")
install.packages("udunits2")
